Do not add pgd script to BASH_ENV
There were 2 problems with this peice of code.

The way it was done was unreliable. If some other script has set
BASH_ENV, then this script wouldn't modify it, so you're never sure
whether or not pgd is loaded and available to you.

Secondly, because I want this to be injected into my environment solely
by my .bashrc file, launching a Bash session with --norc should lead to
a session that does NOT have pgd functions in it. But launching such a
Bash session from a normal session (which has this variable set), also
gets the pgd functions injected into it!