Choose a random port number to start Postgres server
Implemented function pgdChooseServerPortNumber() which tries to find an
unused port number to start the Postgres server on. If the default port
number is available, it uses that, otherwise it picks one from the
unused ports in the range 1025-65535.

The function pgdChooseServerPortNumber() currently works only on
Darwin.

The function pgdGetServerPortNumber() gets the port number from the
running server, so that we don't have to guess or rely on what port
number we used at the time of server start.